+++ packages/javascript/lib/node/os/getHomeDir.ts
@@ -1,0 +1,5 @@
+import { homedir } from 'node:os';
+
+export function getHomeDir(): string {
+	return homedir();
+}