+++ packages/javascript/lib/node/os/getTempDir.ts
@@ -1,0 +1,5 @@
+import { tmpdir } from 'node:os';
+
+export function getTempDir(): string {
+	return tmpdir();
+}